Implementation of a "collapsible" option to hide advanced options.
This change adds a new boolean group option called "collapsible".
When the option is set to True, the group arguments are wrapped in a CollapsiblePane which is folded at start.
This allows to expose avanced options to some users without complexifying too much the interface for average users.
By default, the option set to False to maintain legacy behavior.